# By Marc Alexander (160) and others
# Via Joas Schilling (52) and others
* 'develop' of github.com:imkingdavid/phpbb3: (607 commits)
  [prep-release-3.1.0-a3] Update version to 3.1.0-a4-dev for future development
  [ticket/11880] migration helper file
  [prep-release-3.1.0-a3] Commit changelog for 3.1.0-a3
  [ticket/11880] Move get_schema_steps function to a migrator helper class
  [prep-release-3.1.0-a3] Prepare Changelog
  [prep-release-3.1.0-a3] Add migration for 3.1.0-a3
  [prep-release-3.1.0-a3] Update version to 3.1.0-a3
  [ticket/12163] Use posts_per_page to calculate current page in viewtopic.php
  [develop] Remove left over conflict marker from merge 1a31f03
  [ticket/12162] Set language to C before running diff to ensure its english
  [ticket/11509] Be exact when referencing the branch name.
  [ticket/12141] Disable broken opcache extension on PHP 5.5.7.
  [ticket/12141] PHP extension setup must happen before webserver + PHP-FPM.
  [ticket/12161] Do not use subdirectories in build/save directory.
  [ticket/11985] Do not try to setup PHP extensions on HHVM.
  [ticket/11985] Rename install -> setup. This is more generic and consistent.
  [ticket/11509] Run commit message checker only in one environment.
  [ticket/11509] Contrary to docs, COMMIT_RANGE is not available in PR mode.
  [ticket/12159] Fix codesniffer complaints due to profilefields and passwords
  [ticket/11880] Break up schema changes in the migrator
  ...

Conflicts:
	phpBB/install/schemas/mssql_schema.sql
	tests/test_framework/phpbb_functional_test_case.php

+++ phpBB/phpbb/db/migrator.php
@@ -25,6 +25,9 @@ class migrator
 	/** @var \phpbb\db\tools */
 	protected $db_tools;
 
+	/** @var \phpbb\db\migration\helper */
+	protected $helper;
+
 	/** @var string */
 	protected $table_prefix;
 
@@ -65,11 +68,12 @@ class migrator
 	/**
 	* Constructor of the database migrator
 	*/
-	public function __construct(\phpbb\config\config $config, \phpbb\db\driver\driver $db, \phpbb\db\tools $db_tools, $migrations_table, $phpbb_root_path, $php_ext, $table_prefix, $tools)
+	public function __construct(\phpbb\config\config $config, \phpbb\db\driver\driver $db, \phpbb\db\tools $db_tools, $migrations_table, $phpbb_root_path, $php_ext, $table_prefix, $tools, \phpbb\db\migration\helper $helper)
 	{
 		$this->config = $config;
 		$this->db = $db;
 		$this->db_tools = $db_tools;
+		$this->helper = $helper;
 
 		$this->migrations_table = $migrations_table;
 
@@ -82,6 +86,8 @@ class migrator
 		{
 			$this->tools[$tool->get_name()] = $tool;
 		}
+
+		$this->tools['dbtools'] = $this->db_tools;
 
 		$this->load_migration_state();
 	}
@@ -229,9 +235,12 @@ class migrator
 
 		if (!$state['migration_schema_done'])
 		{
-			$this->last_run_migration['task'] = 'apply_schema_changes';
-			$this->apply_schema_changes($migration->update_schema());
-			$state['migration_schema_done'] = true;
+			$this->last_run_migration['task'] = 'process_schema_step';
+			$steps = $this->helper->get_schema_steps($migration->update_schema());
+			$result = $this->process_data_step($steps, $state['migration_data_state']);
+
+			$state['migration_data_state'] = ($result === true) ? '' : $result;
+			$state['migration_schema_done'] = ($result === true);
 		}
 		else if (!$state['migration_data_done'])
 		{
@@ -329,30 +338,25 @@ class migrator
 
 			$this->set_migration_state($name, $state);
 		}
-		else
+		else if ($state['migration_schema_done'])
 		{
-			$this->apply_schema_changes($migration->revert_schema());
+			$steps = $this->helper->get_schema_steps($migration->revert_schema());
+			$result = $this->process_data_step($steps, $state['migration_data_state']);
 
-			$sql = 'DELETE FROM ' . $this->migrations_table . "
-				WHERE migration_name = '" . $this->db->sql_escape($name) . "'";
-			$this->db->sql_query($sql);
+			$state['migration_data_state'] = ($result === true) ? '' : $result;
+			$state['migration_schema_done'] = ($result === true) ? false : true;
 
-			unset($this->migration_state[$name]);
+			if (!$state['migration_schema_done'])
+			{
+				$sql = 'DELETE FROM ' . $this->migrations_table . "
+					WHERE migration_name = '" . $this->db->sql_escape($name) . "'";
+				$this->db->sql_query($sql);
+
+				unset($this->migration_state[$name]);
+			}
 		}
 
 		return true;
-	}
-
-	/**
-	* Apply schema changes from a migration
-	*
-	* Just calls db_tools->perform_schema_changes
-	*
-	* @param array $schema_changes from migration
-	*/
-	protected function apply_schema_changes($schema_changes)
-	{
-		$this->db_tools->perform_schema_changes($schema_changes);
 	}
 
 	/**
@@ -375,7 +379,7 @@ class migrator
 
 		foreach ($steps as $step_identifier => $step)
 		{
-			$last_result = false;
+			$last_result = 0;
 			if ($state)
 			{
 				// Continue until we reach the step that matches the last step called
@@ -436,7 +440,7 @@ class migrator
 	* @param bool $reverse False to install, True to attempt uninstallation by reversing the call
 	* @return null
 	*/
-	protected function run_step($step, $last_result = false, $reverse = false)
+	protected function run_step($step, $last_result = 0, $reverse = false)
 	{
 		$callable_and_parameters = $this->get_callable_from_step($step, $last_result, $reverse);
 
@@ -459,7 +463,7 @@ class migrator
 	* @param bool $reverse False to install, True to attempt uninstallation by reversing the call
 	* @return array Array with parameters for call_user_func_array(), 0 is the callable, 1 is parameters
 	*/
-	protected function get_callable_from_step(array $step, $last_result = false, $reverse = false)
+	protected function get_callable_from_step(array $step, $last_result = 0, $reverse = false)
 	{
 		$type = $step[0];
 		$parameters = $step[1];
@@ -498,6 +502,7 @@ class migrator
 
 				return $this->get_callable_from_step($step);
 			break;
+
 			case 'custom':
 				if (!is_callable($parameters[0]))
 				{
